WebSockets/tests/submissions/Opera/opening-handshake/002.html

<!doctype html>
<title>WebSockets: valid handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
setTestTimeout(10000, 'got no open event');
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/ws/echo');
  ws.onopen = function(e) {
    
    ws.onclose = function(e) {
      
      assert_equals(e.wasClean, true);
      ws.onclose = t.step_func(function() {assert_unreached()});
      
      setTimeout(t.step_func(function() {t.done();}), 50)
    }
    ws.close();
  }
  ws.onerror = ws.onmessage = ws.onclose = t.step_func(function() {assert_unreached()});
});
</script>
